spark-streaming/python/streamlit/requirements.txt
streamlit==1.54.0
pandas==2.3.3
mysql-connector-python==9.6.0
python-dotenv==1.2.2
